backend/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.2
django-debug-toolbar==3.7.0
django-debug-toolbar-template-profiler==2.1.0
django-extensions==3.2.1
django-taggit==3.0.0
executing==1.1.1
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
oauthlib==3.2.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydotplus==2.0.2
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
wrapt==1.14.1
django-summernote